generic inferer
