--- deeplodocus/core/inference/generic_inferer.py
@@ -1,16 +1,0 @@
-from typing import Tuple
-from typing import Union
-from typing import List
-
-from torch.utils.data import DataLoader
-from torch.nn import Module
-from torch import Tensor
-
-from deeplodocus.data.load.dataset import Dataset
-from deeplodocus.utils.namespace import Namespace
-
-
-class GenericInferer(object):
-
-    def __init__(self):
-        pass
